% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_recharge}
\alias{cal_recharge}
\title{Calculate the surface water recharge using 1D model based on SWAT procedure}
\usage{
cal_recharge(t, R, E0, H2O_SB1, H2O_SB2, params, calibration)
}
\arguments{
\item{t}{A integer vector of simulation day}

\item{R}{A numeric vector of daily precipitation (mm)}

\item{E0}{A numeric vector of daily potential evaporation calculated with Penman monteith method (mm)}

\item{H2O_SB1}{A numeric vector of depth of water in SB1 (initial condition)}

\item{H2O_SB2}{A numeric vector of depth of water in SB2 (initial condition)}

\item{params}{A list for parameters of the model}

\item{calibration}{If true return only the recharge value (for computational efficiency)}
}
\value{
A data.frame with daily surface water recharge
}
\description{
Calculate the surface water recharge using 1D model based on SWAT procedure
}
