% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_sgd}
\alias{cal_sgd}
\title{Calculate the Submarine Groundwater Discharge using an analytical solution for sharp-interface steady state}
\usage{
cal_sgd(GWL, Wrechg, WrechgAve, Pumping, params, calibration)
}
\arguments{
\item{GWL}{A numeric vector of daily groundwater level (mm)}

\item{Wrechg}{A numeric vector of daily surface water recharge (mm) calculated from `cal_recharge()`.}

\item{WrechgAve}{A numeric vector of moving average of `Wrechg`}

\item{Pumping}{A numeric vector of daily pumping rate}

\item{params}{A list for parameters of the model}

\item{calibration}{If true return only the water level value (for computational efficiency)}

\item{H2O_SB2}{A numeric vector of depth of water in SB2 (initial condition)}
}
\value{
A data.frame with daily SGD
}
\description{
Calculate the Submarine Groundwater Discharge using an analytical solution for sharp-interface steady state
}
