% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{change_unknown_pumping}
\alias{change_unknown_pumping}
\title{A function to change the unknown pumping rate to a preferred value.}
\usage{
change_unknown_pumping(x, input_df, yearly_df, pumping_df)
}
\arguments{
\item{x}{A numeric value of the preferred pumping rate}

\item{input_df}{A data.frame of the input data. See \link{estimate_sgd} for details.}

\item{yearly_df}{A data.frame of the yearly pumping data. (2 columns: year, pumping)}

\item{pumping_df}{A data.frame of the pumping data}
}
\value{
A data.frame of the input data with the preferred pumping rate
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
