% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{json_to_paramter_list}
\alias{json_to_paramter_list}
\alias{json_to_parameter_list}
\title{Convert a JSON file to a list of parameters}
\usage{
json_to_parameter_list(json_file)
}
\arguments{
\item{json_file}{A JSON file containing the parameters}
}
\value{
A list of parameters (invisible)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function converts a JSON file to a list of parameters.
}
