% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{model_diagnose}
\alias{model_diagnose}
\alias{model_diagnosis}
\title{Output the model diagnose information including estimated groundwater level, hn, xn, recharge and average recharge.}
\usage{
model_diagnosis(result, obs_data, start_date)
}
\arguments{
\item{result}{The model estimated results. Generated by \code{\link{estimate_sgd}}.}

\item{obs_data}{The data.frame used as \code{inputData} in \code{\link{estimate_sgd}}.}

\item{start_date}{A valid string that can be transformed into data format using \code{ymd}.}
}
\value{
A \code{ggplot} object.
}
\description{
Output the model diagnose information including estimated groundwater level, hn, xn, recharge and average recharge.
}
