% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moving_average}
\alias{moving_average}
\title{Return a numeric vector of moving average with leading NAs filled with first average value}
\usage{
moving_average(x, windowSize)
}
\arguments{
\item{x}{A numeric vector with no NAs}

\item{windowSize}{A integer of window size}
}
\value{
A numeric vector of moving average with leading NAs filled with first average value
}
\description{
Return a numeric vector of moving average with leading NAs filled with first average value
}
