% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{parameter_list_to_vector}
\alias{parameter_list_to_vector}
\title{Convert a list of parameters to a vector of parameters}
\usage{
parameter_list_to_vector(parameter_list)
}
\arguments{
\item{parameter_list}{A list of parameters}
}
\value{
A named vector of parameters
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function converts a list of parameters to a vector of parameters.
}
