% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassExports.R
\name{Parameters}
\alias{Parameters}
\title{Create a Bucket Object from the Bucket C++ Class}
\arguments{
\item{layer}{Soil layer number in bucket (-). In this model we only have 2 layers (1 and 2).}
}
\value{
A `Bucket` object from the _C++_ Bucket Class.
}
\description{
Allows for the creation of a Bucket Object in _C++_ from _R_
using the _C++_ Bucket class.
}
\examples{
##################
## Constructor

# Construct new Bucket object called "bucket1"

bucketParams <- list(layer = 1, 
                     WP = 100, 
                     z = 200, 
                     FCmm = 300, 
                     SAT = 400, 
                     Swres = 500, 
                     Ksat = 600, 
                     n = 0.5, 
                     phi_soil = 0.3, 
                     Y = 0.2, 
                     Z = 0.1)
bucket1 <- new(Bucket, bucketParams)

##################
## Getters

bucket1$get_WP()
bucket1$get_z()
bucket1$get_FCmm()
bucket1$get_SAT()
bucket1$get_Swres()
bucket1$get_Ksat()
bucket1$get_n()
bucket1$get_phi_soil()
bucket1$get_Y()
bucket1$get_Z()
}
