% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{paramter_vec_to_list}
\alias{paramter_vec_to_list}
\alias{parameter_vec_to_list}
\title{Convert a vector of parameters to a list of parameters (useful for calibrating parameters)}
\usage{
parameter_vec_to_list(parameter_vec, prameter_skeleton)
}
\arguments{
\item{parameter_vec}{A vector of parameters}

\item{prameter_skeleton}{A list of parameters to use as a skeleton for the new list}
}
\value{
A list of parameters (invisible)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function converts a vector of parameters to a list of parameters.
}
