% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interface.R
\name{create_sgd_model}
\alias{create_sgd_model}
\alias{R_model_interface}
\title{Gernerate a SGD model in R}
\usage{
create_sgd_model(input_data, calibratable_parameters, const_paramters)
}
\arguments{
\item{input_data}{A data frame containing the input data for the model}

\item{calibratable_parameters}{A list of calibratable parameters}

\item{const_paramters}{A list of constant parameters}
}
\value{
A \code{Model} object from the \emph{C++} Model Class.
}
\description{
This function allows for the creation of a \code{Model} object in C++ from R.
}
\details{
This function allows for the creation of a \code{Model} object in C++ from _R.
\itemize{
\item The \code{input_data} should be a data.frame containing the required columns for the model (R, E0, Pumping, GWL, H2O_SB1, H2O_SB2).
\itemize{
\item \code{R} is a numeric vector of the daily precipitation (mm).
\item \code{E0} is a numeric vector of the daily potential evapotranspiration (mm).
\item \code{Pumping} is a numeric vector of the daily pumping rate (m3/d).
\item \code{GWL} is a numeric vector of the initial groundwater level (mm), Only the first value is used as the initial value for the SB1.
\item \code{H2O_SB1} is a numeric vector of the initial water level in the first soil bucket (mm). Only the first value is used as the initial value for the SB1.
\item \code{H2O_SB2} is a numeric vector of the inital water level in the second soil bucket (mm). Only the first value is used as the initial value for the SB2.
}
\item \code{calibratable_parameters} should be a list of calibratable parameters. It's recommended to use \code{\link{json_to_paramter_list}} to create this list.
\item \code{const_paramters} should be a list of constant parameters. It's recommended to use \code{\link{json_to_paramter_list}} to create this list.
}
}
