% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDr-package.R
\docType{package}
\name{SGDr-package}
\alias{SGDr}
\alias{SGDr-package}
\title{SGDr: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hxfan1227.github.io/SGDr/}
}

}
