% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{SGDr_example}
\alias{SGDr_example}
\title{Get path to \code{SGDr} example}
\usage{
SGDr_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
SGDr comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
SGDr_example()
SGDr_example("test_data.csv")
}
