% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{summary.SGD_ESTIMATION_DF}
\alias{summary.SGD_ESTIMATION_DF}
\title{Summary the SGD estimation results.}
\usage{
\method{summary}{SGD_ESTIMATION_DF}(object, base_date, ...)
}
\arguments{
\item{object}{An object of class \code{sgdEstimation}. Created by \code{\link{estimate_sgd}}.}

\item{base_date}{A valid string that can be transformed into data format using \code{\link[lubridate]{ymd}}.}

\item{...}{other arguments not used by this method}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
