% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_sgd_c}
\alias{cal_sgd_c}
\title{Calculate the Submarine Groundwater Discharge using an analytical solution for sharp-interface steady state (for calibration)}
\usage{
cal_sgd_c(GWL, Wrechg, WrechgAve, Pumping, params, consts, sgd = FALSE)
}
\arguments{
\item{GWL}{A numeric vector of daily groundwater level (mm)}

\item{Wrechg}{A numeric vector of daily surface water recharge (mm) calculated from `cal_recharge()`.}

\item{WrechgAve}{A numeric vector of moving average of `Wrechg`}

\item{Pumping}{A numeric vector of daily pumping rate}

\item{params}{A list for parameters of the model}

\item{consts}{A list for constants of the model}

\item{sgd}{if TRUE return SDG time series. The default is FALSE.}

\item{H2O_SB2}{A numeric vector of depth of water in SB2 (initial condition)}
}
\value{
A vector of daiyly SGD or water level.
}
\description{
Calculate the Submarine Groundwater Discharge using an analytical solution for sharp-interface steady state (for calibration)
}
