% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{estimate_sgd_from_pars}
\alias{estimate_sgd_from_pars}
\title{A wrapper function to call the model and return result for calibration.}
\usage{
estimate_sgd_from_pars(
  pars,
  parnames = c(params_to_calibrate, "nw"),
  parset = params_to_calibrate,
  input = test_data,
  warm_up = 1500,
  default_pars = preferred_params,
  skeleton = params_skeleton,
  const_par_list = const_params_list
)
}
\arguments{
\item{pars}{A numeric vector of parameters to calibrate}

\item{parnames}{A character vector of names of the parameters}

\item{parset}{A character vector of names of the to-be-calibrated parameters}

\item{input}{A data.table of the input data}

\item{warm_up}{A numeric value of the warm-up period}

\item{default_pars}{A named numeric vector of the default parameters}

\item{skeleton}{A named list of the parameter skeleton}

\item{const_par_list}{A named list of the constant parameters}
}
\value{
A SGD_ESTIMATION_DF class
}
\description{
A wrapper function to call the model and return result for calibration.
}
\seealso{
\code{\link[=estimate_sgd]{estimate_sgd()}}
}
