% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{estimate_sgd_from_pars}
\alias{estimate_sgd_from_pars}
\title{A wrapper function to call the model and return result for calibration.}
\usage{
estimate_sgd_from_pars(
  pars,
  parnames,
  parset,
  skeleton,
  inputDf,
  yearlyPumping,
  pumpingDf,
  ...
)
}
\arguments{
\item{pars}{A numeric vector of parameters to calibrate.}

\item{parnames}{A character vector of names of the parameters.}

\item{parset}{A character vector of names of the to-be-calibrated parameters.}

\item{skeleton}{A named list of the parameter skeleton.}

\item{inputDf}{A data.frame of the input data. See \link{estimate_sgd} for details.}

\item{yearlyPumping}{A data.frame of the yearly pumping data. (2 columns: year, pumping)}

\item{pumpingDf}{A data.frame of the pumping data.}

\item{...}{Additional arguments passed to the model function}
}
\value{
A SGD_ESTIMATION_DF class
}
\description{
A wrapper function to call the model and return result for calibration.
}
\seealso{
\code{\link{estimate_sgd}}
}
