% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{json_to_paramter_list}
\alias{json_to_paramter_list}
\title{Convert a JSON file to a list of parameters}
\usage{
json_to_paramter_list(json_file)
}
\arguments{
\item{json_file}{A JSON file containing the parameters}
}
\value{
A list of parameters (invisible)
}
\description{
This function converts a JSON file to a list of parameters.
}
