% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{model_diagnose}
\alias{model_diagnose}
\title{Output the model diagnose information including estimated groundwater level, hn, xn, recharge and average recharge.}
\usage{
model_diagnose(result, obs_data, start_date)
}
\arguments{
\item{result}{The model estimated results. Generated by \code{\link{estimate_sgd}}.}

\item{obs_data}{A data.frame of daily observed groundwater level data. See Details.}

\item{start_date}{A valid string that can be transformed into data format using \code{\link[lubridate]{ymd}}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Output the model diagnose information including estimated groundwater level, hn, xn, recharge and average recharge.
}
\details{
The \code{obs_data} is a data.frame containing the following 2 columns:
\itemize{
\item \code{date}   Observation date
\item \code{wl}     Daily observed groundwater level (m)
}
}
