% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassExports.R
\name{Parameters}
\alias{Parameters}
\title{Create a Parameters Object from the Parameters C++ Class}
\arguments{
\item{name}{Name of Parameters}

\item{age}{Age of Parameters}

\item{male}{Is Parameters a Male?}
}
\value{
A \code{Parameters} object from the \emph{C++} Parameters Class.
}
\description{
Allows for the creation of a Parameters Object in \emph{C++} from \emph{R}
using the \emph{C++} Parameters class.
}
\examples{
##################
## Constructor

# Construct new Parameters object called "ben"
ben = new(Parameters, name = "Ben", age = 26, male = TRUE)

##################
## Getters

ben$LikesBlue()

ben$GetAge()

ben$IsMale()

ben$GetName()

ben$GetFavoriteNumbers()
}
