% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{plot.SGD_ESTIMATION_DF}
\alias{plot.SGD_ESTIMATION_DF}
\title{Plot the SGD estimation results}
\usage{
\method{plot}{SGD_ESTIMATION_DF}(
  x,
  y,
  vars = c("wl", "SGD", "Wrechg"),
  base_date,
  type = c("pred", "comp", "input"),
  obs_x = "date",
  obs_y = "wl",
  xbreaks = 10,
  ybreaks = 5,
  maxRange = 600,
  coeff = 0.3,
  names = c("wl", "SGD", "hn", "xn", "Wrechg", "WrechgAve"),
  colors = c(obs = "red", precp = "blue", pumping = "green", est = "blue"),
  color_labels = c(obs = "Observed GWL", precp = "Precipitation", pumping = "Pumping",
    est = "Estimated GWL"),
  labels = c("WL~(m~AHD)", "SGD~(m^3/d)", "h[n]~(m~AHD)", "x[n]~(m)", "w[recharge]~(mm)",
    "w[recharge[ave]]~(mm)"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{sgdEstimation}. Created by \code{\link{estimate_sgd}}.}

\item{y}{A data.frame of daily observed groundwater level data. See Details.}

\item{base_date}{A valid string that can be transformed into data format using \code{\link[lubridate]{ymd}}.}

\item{type}{A character string indicating the type of plot. Can be one of 'pred', 'comp', 'input'.}

\item{obs_x}{A character string indicating the column name of the date in the observed data.}

\item{obs_y}{A character string indicating the column name of the groundwater level in the observed data.}

\item{xbreaks}{An integer indicating the number of breaks on the x-axis.}

\item{ybreaks}{An integer indicating the number of breaks on the y-axis.}

\item{maxRange}{A numeric indicating the maximum range of the first axis (groundwater level).}

\item{coeff}{A numeric indicating the shrink coefficient of the precipitation.}

\item{names}{A character vector of the variable names to plot.}

\item{colors}{A named character vector of the colors to use}

\item{color_labels}{A named character vector of the labels for the colors.}

\item{labels}{A character vector of the variable labels to plot.}

\item{...}{Additional arguments passed to the plot function. (currently not implemented)}
}
\value{
Invisibly returns the original estimated results.
}
\description{
Plot the SGD estimation results
}
