% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prepare_warm_up}
\alias{prepare_warm_up}
\title{Prepare the warm-up data}
\usage{
prepare_warm_up(data, length)
}
\arguments{
\item{data}{A data.frame containing the input data for the model.}

\item{length}{An integer indicating the length of the warm-up period (days).}
}
\value{
A data.frame containing the input data for the model with the warm-up period.
}
\description{
Prepare the warm-up data
}
