% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{print.SGD_ESTIMATION_DF}
\alias{print.SGD_ESTIMATION_DF}
\title{Print the SGD estimation results.}
\usage{
\method{print}{SGD_ESTIMATION_DF}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SGD_ESTIMATION_DF}. Created by \code{\link{estimate_sgd}}.}

\item{...}{other arguments not used by this method}
}
\description{
Print the SGD estimation results.
}
