% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDr-package.R
\docType{package}
\name{SGDr-package}
\alias{SGDr}
\alias{SGDr-package}
\title{SGDr: Submarine Groundwater Discharge (SGD) calculation in R}
\description{
Estimate SGD using 1-D bucket model and Strack (1976) analytical solution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hxfan1227.github.io/SGDr/}
}

}
\author{
\strong{Maintainer}: Hongxiang Fan \email{hxfan@niglas.ac.cn} (\href{https://orcid.org/0000-0001-7626-1344}{ORCID})

Authors:
\itemize{
  \item Adrian Werner \email{adrian.werner@flinders.edu.au} (\href{https://orcid.org/0000-0002-1190-1301}{ORCID}) [contributor]
  \item Silvia Cristina Solórzano-Rivas \email{cristina.solorzano@flinders.edu.au} (\href{https://orcid.org/0000-0003-4436-5813}{ORCID}) [contributor]
}

}
