% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{summary}
\alias{summary}
\alias{summary.SGD_ESTIMATION_DF}
\title{Summary the SGD estimation results.}
\usage{
summary(object, ...)

\method{summary}{SGD_ESTIMATION_DF}(object, base_date, ...)
}
\arguments{
\item{object}{An object of class \code{sgdEstimation}. Created by \code{\link{estimate_sgd}}.}
}
\description{
Summary the SGD estimation results.
}
